[REWORK]: Errors
