+++ apps/ledger-live-desktop/src/newArch/WalletSync/hooks/walletSync.hooks.ts
@@ -1,48 +1,11 @@
-import {
-  trustchainSelector,
-  memberCredentialsSelector,
-  setTrustchain,
-  resetTrustchainStore,
-} from "@ledgerhq/trustchain/store";
-import { Trustchain, MemberCredentials } from "@ledgerhq/trustchain/types";
-import { useQuery } from "@tanstack/react-query";
-import { useDispatch, useSelector } from "react-redux";
-import { useTrustchainSdk } from "./useTrustchainSdk";
-import { ErrorType, QueryKey } from "./type.hooks";
+import { resetTrustchainStore } from "@ledgerhq/trustchain/store";
+import { useDispatch } from "react-redux";
+import { ErrorType } from "./type.hooks";
 import { setFlow } from "~/renderer/actions/walletSync";
 import { Flow, Step } from "~/renderer/reducers/walletSync";
-import { TrustchainEjected } from "@ledgerhq/trustchain/errors";
-
-export const useRestoreTrustchain = () => {
-  const dispatch = useDispatch();
-  const sdk = useTrustchainSdk();
-  const memberCredentials = useSelector(memberCredentialsSelector);
-
-  const trustchain = useSelector(trustchainSelector);
-
-  async function restoreTrustchain() {
-    const newTrustchain = await sdk.restoreTrustchain(
-      trustchain as Trustchain,
-      memberCredentials as MemberCredentials,
-    );
-
-    dispatch(setTrustchain(newTrustchain));
-
-    return newTrustchain;
-  }
-
-  const restore = useQuery({
-    enabled: false,
-    queryKey: [QueryKey.restoreTrustchain],
-    queryFn: restoreTrustchain,
-  });
-
-  return restore;
-};
+import { TrustchainEjected, TrustchainNotAllowed } from "@ledgerhq/trustchain/errors";
 
 export const useLifeCycle = () => {
-  const restoreTrustchain = useRestoreTrustchain();
-
   const dispatch = useDispatch();
 
   function reset() {
@@ -51,17 +14,15 @@ export const useLifeCycle = () => {
   }
 
   const includesErrorActions: { [key: string]: () => void } = {
-    [ErrorType.NOT_MEMBER]: () => reset(),
-    [ErrorType.NO_PERMISSION]: () => reset(),
-    [ErrorType.NO_PERMISSION_FOR_TRUSTCHAIN]: () => restoreTrustchain.refetch(),
     [ErrorType.NO_TRUSTCHAIN]: () => reset(),
   };
 
   function handleError(error: Error) {
-    console.error("GetMember : " + error);
-    if (error instanceof TrustchainEjected) {
-      reset();
-    }
+    console.error("GetMember :" + error);
+
+    if (error instanceof TrustchainEjected) reset();
+    if (error instanceof TrustchainNotAllowed) reset();
+
     const errorToHandle = Object.entries(includesErrorActions).find(([err, _action]) =>
       error.message.includes(err),
     );
